Restructured project; added cover image transition.
