+++ Components/Radio/RadioError.swift
@@ -1,0 +1,4 @@
+enum RadioError: Error {
+    case metadataError
+    case playbackError
+}
