+++ Components/Radio/RadioMetadataFetcher.swift
@@ -1,17 +1,6 @@
 import Foundation
 import MediaPlayer
 
-struct RadioMetadata: Codable {
-    var artist : String
-    var cover : URL
-    var duration : Double
-    var end_at : String
-    var id : Int
-    var next_track : String
-    var started_at : String
-    var title : String
-}
-
 @MainActor
 class RadioMetadataFetcher: ObservableObject {
     private let nowPlayingInfoCenter = MPNowPlayingInfoCenter.default()
