+++ Components/Radio/RadioStatus.swift
@@ -1,0 +1,5 @@
+enum RadioStatus: String {
+    case stopped
+    case buffering
+    case playing
+}
